pyproject.toml
[project]
name = "quickchart-mcp"
version = "0.1.0"
description = """
An MCP server implementation for QuickChart.io, allowing you to generate and download charts 
directly through Claude Desktop and other MCP clients.
"""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.9.0",
    "python-dotenv>=1.1.0",
    "quickchart-io>=2.0.0",
]
